101-print_number.c

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -38,6 +38,6 @@ for (j = 0; j < i; j++)
 {
 exp = exp * 10;
 }
-_putchar(((abs / exp) % 10) + '0');
+_putchar((((unsigned int)abs / exp) % 10) + '0');
 }
 }
